# Clean up `influx_inspect export` command and improve performance