@mark-rushakoff 2016-12-19
w needs to be typed as an IO Writer so the gzip writer can be conditionally
assigned to w later.

In fact I need to double check whether it performs better with the buffered
writer in front of the gzip writer. I suspect it will.

On Mon, Dec 19, 2016 at 4:30 AM Edd Robinson <notifications@github.com>
wrote:

> *@e-dard* approved this pull request.
>
>
>
>
> Just a small nit, otherwise LGTM 👍
> ------------------------------
>
>
>
>
> In cmd/influx_inspect/export/export.go
> <https://github.com/influxdata/influxdb/pull/7741#pullrequestreview-13542021>
> :
>
>
> >  	if err != nil {
>
>  		return err
>
>  	}
>
> -	defer w.Close()
>
> +	defer f.Close()
>
> +
>
> +	// Because calling (*os.File).Write is relatively expensive,
>
> +	// and we don't *need* to sync to disk on every written line of export,
>
> +	// use a sized buffered writer so that we only sync the file every megabyte.
>
> +	bw := bufio.NewWriterSize(f, 1024*1024)
>
>
>
> Can this not just be w := bufio.... and then we don't need the var w = bw
> below?
>
>
>
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/influxdata/influxdb/pull/7741#pullrequestreview-13542021>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AARvV1bBrt_w4KrcOlX4z1Bt85i1rxuaks5rJnjtgaJpZM4LQC9j>
> .
>
>
>
>
>
>
>
>
>
>
>
>
>
>
>
>
>
>
>