@mark-rushakoff 2016-12-19
Using my earlier test case of a 450MB TSM file that, when time-filtered, exports to a 21GB plain text file or a 650MB gzipped text file, I found:

```
time influx_inspect export -datadir . -waldir . -start '2016-12-13T10:59:49Z' -compress -out ./s4.2.txt.gz

# 1.1.1
real	5m7.206s
user	5m8.761s
sys	0m7.027s

# buf -> gzw -> f
real	3m53.893s
user	3m51.619s
sys	0m5.935s

# gzw -> buf -> f (behavior as committed in PR)
real	3m52.385s
user	3m54.782s
sys	0m1.984s
```

So the real time is about the same whether the buffer writer is before or after the gzip writer, but putting the gzip writer before the buffered writer seems to result in fewer file writes, which seems like the better win.