@jwilder 2016-12-19
Nice speedup!